+++ tilings/strategies/deflation.py
@@ -71,21 +71,17 @@ class DeflationStrategy(Strategy[Tiling, GriddedPerm]):
         self.sum_deflate = sum_deflate
         super().__init__()
 
-    @staticmethod
-    def can_be_equivalent() -> bool:
+    def can_be_equivalent(self) -> bool:
         return False
 
-    @staticmethod
-    def is_two_way(comb_class: Tiling) -> bool:
+    def is_two_way(self, comb_class: Tiling) -> bool:
         return False
 
-    @staticmethod
-    def is_reversible(comb_class: Tiling) -> bool:
+    def is_reversible(self, comb_class: Tiling) -> bool:
         return False
 
-    @staticmethod
     def shifts(
-        comb_class: Tiling, children: Optional[Tuple[Tiling, ...]] = None
+        self, comb_class: Tiling, children: Optional[Tuple[Tiling, ...]] = None
     ) -> Tuple[int, ...]:
         return (0, 0)