+++ tilings/strategies/factor.py
@@ -199,11 +199,10 @@ class Interleaving(CartesianProduct[Tiling, GriddedPerm]):
         self.insertion_constructor = insertion_constructor
 
     @staticmethod
-    def is_equivalence() -> bool:
+    def is_equivalence(is_empty: Optional[Callable[[Tiling], bool]] = None) -> bool:
         return False
 
-    @staticmethod
-    def get_equation(lhs_func: Function, rhs_funcs: Tuple[Function, ...]) -> Eq:
+    def get_equation(self, lhs_func: Function, rhs_funcs: Tuple[Function, ...]) -> Eq:
         raise NotImplementedError
 
     def get_terms(
@@ -276,10 +275,10 @@ class FactorWithInterleavingStrategy(FactorStrategy):
         )
         return f"{self.__class__.__name__}({args})"
 
-    def is_two_way(self, comb_class: Tiling) -> bool:  # type: ignore
+    def is_two_way(self, comb_class: Tiling) -> bool:  # pylint: disable=W0221
         return self.is_reversible(comb_class)
 
-    def is_reversible(self, comb_class: Tiling) -> bool:  # type: ignore
+    def is_reversible(self, comb_class: Tiling) -> bool:  # pylint: disable=W0221
         return not bool(self.assumptions_to_add(comb_class))
 
     def formal_step(self) -> str:
@@ -613,8 +612,9 @@ class FactorFactory(StrategyFactory[Tiling]):
             FactorWithInterleavingStrategy.interleaving_rows_and_cols(components)
         )
         if interleaving:
+            # pylint: disable=E1123
             return cast(
-                FactorStrategy,
+                FactorWithInterleavingStrategy,
                 self.factor_class(
                     components,
                     ignore_parent=self.ignore_parent,