+++ tilings/tiling.py
@@ -6,6 +6,7 @@ from functools import reduce
 from itertools import chain, filterfalse, product
 from operator import mul, xor
 from typing import (
+    Any,
     Callable,
     Dict,
     FrozenSet,
@@ -1514,7 +1515,7 @@ class Tiling(CombinatorialClass):
             if len(gp) == length:
                 yield gp
 
-    def initial_conditions(self, check: int = 6) -> List[sympy.Expr]:
+    def initial_conditions(self, check: int = 6) -> List[Any]:
         """
         Returns a list with the initial conditions to size `check` of the
         CombinatorialClass.
@@ -1834,7 +1835,7 @@ class Tiling(CombinatorialClass):
             requirements=[[GriddedPerm((0,), ((i, p[i]),))] for i in range(len(p))]
         )
 
-    def get_genf(self, *args, **kwargs) -> sympy.Expr:
+    def get_genf(self, *args, **kwargs) -> Any:
         # pylint: disable=import-outside-toplevel
         if self.is_empty():
             return sympy.sympify(0)