+++ CHANGELOG.md
@@ -51,6 +51,11 @@ swapped around a fusable row or column.
 - `RearrangeAssumptionFactory` will ignore component assumptions
 - `GriddedPermReduction.minimal_reqs` was removing requirements if they 
   were duplicates.
+- `RequirementPlacement` algorithm didn't minimise obstructions correctly when
+  placing size 2 or higher gridded perms.
+- added missing condition in `MonotoneSlidingFactory` for consecutive 
+  values. Previous rules failing this condition will now raise 
+  `StrategyDoesNotApply` if it fails this condition.
 
 ### Changed
 - `TileScopePack.make_tracked` will add the appropriate tracking methods for